+++ src/components/header/Header.tsx
@@ -1,12 +1,13 @@
 import { Ghost, Waypoints } from 'lucide-react';
-import { name } from '@package';
+import { name, version } from '@package';
 import Link from 'next/link';
+import { capitalize } from '@/util/capitalize';
 
 export function Header({ children, ...props }: React.ComponentProps<'header'>) {
   return (
     <header
       {...props}
-      className='py-4 px-default flex items-center w-full border-b border-gray-600 justify-between z-20 bg-background-light'>
+      className='py-4 px-default flex items-center w-full border-b border-background-light-border justify-between z-20 bg-background-light'>
       <Logo />
       <div className='flex items-center gap-4'>{children}</div>
     </header>
@@ -14,6 +15,8 @@ export function Header({ children, ...props }: React.ComponentProps<'header'>) {
 }
 
 export function Logo() {
+  const versionString = version.split('-').at(1);
+
   return (
     <Link
       href='/'
@@ -27,7 +30,7 @@ export function Logo() {
           className='font-semibold tracking-wider text-[10px]'>
           {name}
         </h2>
-        <small className='text-accent'>Beta</small>
+        <small className='text-accent'>{versionString && capitalize(versionString)}</small>
       </div>
     </Link>
   );