+++ src/features/events/schemas/eventSchema.ts
@@ -21,7 +21,8 @@ export const eventInstanceSchema = z.object({
   created_at: z.date().optional(),
   ended_at: z.date().optional(),
   location: z.object().optional(),
-  position: z.object(),
+  position: z.object().optional(),
+  position_accuracy: z.number().optional(),
 });
 
 export const eventSchema = z